+++ include/treelike/renderer/circle.h
@@ -7,18 +7,18 @@ namespace renderer
     class circle : public node
     {
     public:
-        CREATE_H( circle, float radius, float segments = -1.0F );
+        CREATE_H( circle, float radius, int segments = -1 );
         virtual ~circle( );
-        bool init( float radius, float segments = -1.0F );
+        bool init( float radius, int segments = -1 );
         virtual void render( ) override;
     public:
         void set_radius( float value );
         float get_radius( );
-        void set_segments( float value );
-        float get_segments( );
+        void set_segments( int value );
+        int get_segments( );
     protected:
         float _radius = 0.0F;
-        float _segments = -1.0F;
+        int _segments = -1;
     };
 }
 }
